Move filter_kwargs_to_str() to filter_support.exceptions
